pyproject.toml
[project]
name = "friday"
version = "0.2.0"
description = "Friday multi-agent business operating system"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110.0",
  "pydantic>=2.6.0",
  "uvicorn>=0.29.0",
  "anthropic>=0.25.0",
  "openai>=1.0.0",
  "python-multipart>=0.0.9",  # required for POST /upload (file uploads)
]

[project.optional-dependencies]
# Phase 3 production dependencies:
#   pip install -e ".[phase3]"
# Then set FRIDAY_MEMORY_DATABASE_URL=postgresql://... to enable Postgres+pgvector
# and FRIDAY_AUDIT_DATABASE_URL=postgresql://... for the audit store.
phase3 = [
  "psycopg[binary]>=3.1.0",   # Postgres driver (also used by pgvector path)
  "temporalio>=1.7.0",         # Durable workflow engine (optional; set FRIDAY_WORKFLOW_ENGINE=temporal)
]
# Phase 4 multi-modal file extraction:
#   pip install -e ".[phase4]"
phase4 = [
  "pypdf>=4.0.0",              # PDF text extraction for POST /upload
  "python-docx>=1.1.0",        # DOCX text extraction for POST /upload
]
# Document generation (Word, PowerPoint, Excel, PDF):
#   pip install -e ".[docgen]"
docgen = [
  "python-docx>=1.1.0",        # Word document generation
  "python-pptx>=0.6.23",       # PowerPoint generation
  "openpyxl>=3.1.2",           # Excel generation
  "Jinja2>=3.1.3",             # Template rendering
  "weasyprint>=62.0",          # HTML→PDF (optional, graceful fallback)
]
# Analytics and charting:
#   pip install -e ".[analytics]"
analytics = [
  "plotly>=5.22.0",            # Interactive charts
  "kaleido>=0.2.1",            # Plotly static image export
]
# All integrations:
#   pip install -e ".[integrations]"
integrations = [
  "google-api-python-client>=2.120.0",
  "google-auth-oauthlib>=1.2.0",
  "google-auth-httplib2>=0.2.0",
  "slack-sdk>=3.27.0",
  "msal>=1.28.0",              # Microsoft OAuth (Outlook, Calendar)
  "jira>=3.8.0",
  "atlassian-python-api>=3.41.0",
  "notion-client>=2.2.0",
  "simple-salesforce>=1.12.0",
  "hubspot-api-client>=9.0.0",
]
# Everything:
#   pip install -e ".[all]"
all = [
  "friday[docgen,analytics,integrations,phase3,phase4]",
]
llm-anthropic = [
  "anthropic>=0.25.0",
]
llm-openai = [
  "openai>=1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
